+++ src/main/gen/org/intellij/sdk/language/DependencyLanguageLexer.java
@@ -25,6 +25,7 @@ public class DependencyLanguageLexer implements FlexLexer {
 
   /** lexical states */
   public static final int YYINITIAL = 0;
+  public static final int DEPENDENCY = 2;
 
   /**
    * ZZ_LEXSTATE[l] is the state in the DFA for the lexical state l
@@ -33,7 +34,7 @@ public class DependencyLanguageLexer implements FlexLexer {
    * l is of the form l = 2*k, k a non negative integer
    */
   private static final int ZZ_LEXSTATE[] = { 
-     0, 1
+     0,  1,  2, 3
   };
 
   /** 
@@ -55,9 +56,9 @@ public class DependencyLanguageLexer implements FlexLexer {
 
   /* The ZZ_CMAP_A table has 256 entries */
   static final char ZZ_CMAP_A[] = zzUnpackCMap(
-    "\11\0\1\3\1\6\1\0\1\3\1\5\22\0\1\3\1\13\1\0\1\4\1\23\3\0\1\16\1\17\1\0\1\2"+
-    "\1\15\1\22\1\2\1\0\12\1\1\7\1\0\1\12\1\11\1\10\2\0\32\1\1\20\1\0\1\21\1\0"+
-    "\1\2\1\0\32\1\1\24\1\14\1\25\1\2\201\0");
+    "\11\0\1\4\1\7\1\0\1\4\1\6\22\0\1\4\1\14\1\0\1\5\1\23\3\0\1\17\1\20\1\0\1\3"+
+    "\1\16\1\10\1\3\1\0\12\1\1\2\1\0\1\13\1\12\1\11\2\0\32\1\1\21\1\0\1\22\1\0"+
+    "\1\3\1\0\32\1\1\24\1\15\1\25\1\3\201\0");
 
   /** 
    * Translates DFA states to action switch labels.
@@ -65,12 +66,12 @@ public class DependencyLanguageLexer implements FlexLexer {
   private static final int [] ZZ_ACTION = zzUnpackAction();
 
   private static final String ZZ_ACTION_PACKED_0 =
-    "\2\0\1\1\1\2\1\3\1\4\1\5\1\6\1\7"+
-    "\1\10\1\11\1\12\1\13\1\14\1\15\1\16\1\1"+
-    "\1\17\1\20\1\6\1\21\3\17";
+    "\4\0\1\1\1\2\1\3\1\4\1\1\1\5\1\6"+
+    "\1\7\1\10\1\11\1\12\1\13\1\14\1\15\1\16"+
+    "\1\17\1\20\2\6\1\21\1\6\3\17";
 
   private static int [] zzUnpackAction() {
-    int [] result = new int[24];
+    int [] result = new int[28];
     int offset = 0;
     offset = zzUnpackAction(ZZ_ACTION_PACKED_0, offset, result);
     return result;
@@ -96,11 +97,12 @@ public class DependencyLanguageLexer implements FlexLexer {
 
   private static final String ZZ_ROWMAP_PACKED_0 =
     "\0\0\0\26\0\54\0\102\0\130\0\156\0\204\0\232"+
-    "\0\260\0\232\0\54\0\54\0\54\0\54\0\54\0\54"+
-    "\0\306\0\334\0\362\0\54\0\54\0\u0108\0\u011e\0\54";
+    "\0\260\0\306\0\334\0\362\0\334\0\130\0\130\0\130"+
+    "\0\130\0\130\0\130\0\u0108\0\u011e\0\306\0\362\0\130"+
+    "\0\130\0\u0134\0\u014a\0\130";
 
   private static int [] zzUnpackRowMap() {
-    int [] result = new int[24];
+    int [] result = new int[28];
     int offset = 0;
     offset = zzUnpackRowMap(ZZ_ROWMAP_PACKED_0, offset, result);
     return result;
@@ -123,20 +125,24 @@ public class DependencyLanguageLexer implements FlexLexer {
   private static final int [] ZZ_TRANS = zzUnpackTrans();
 
   private static final String ZZ_TRANS_PACKED_0 =
-    "\1\3\1\4\1\3\1\5\1\6\1\3\1\5\1\3"+
-    "\1\7\1\10\1\11\1\12\1\13\1\14\1\15\1\16"+
-    "\1\17\1\20\1\21\1\22\3\3\1\4\1\3\1\5"+
-    "\1\23\1\3\1\5\1\3\1\7\1\10\1\11\1\12"+
-    "\1\13\1\14\1\15\1\16\1\17\1\20\1\21\1\22"+
-    "\2\3\27\0\2\4\4\0\1\4\12\0\1\4\6\0"+
-    "\1\5\2\0\1\5\17\0\5\6\2\0\17\6\10\0"+
-    "\2\24\25\0\1\24\25\0\2\24\35\0\1\25\4\0"+
-    "\1\26\22\0\1\27\1\0\5\23\2\0\17\23\1\0"+
-    "\1\26\5\0\1\26\12\0\1\26\2\0\1\30\1\0"+
-    "\1\26\23\0\1\30";
+    "\1\5\1\6\2\5\1\7\1\10\1\5\1\7\1\11"+
+    "\1\12\1\13\1\14\1\15\1\16\1\17\1\20\1\21"+
+    "\1\22\1\23\1\24\3\5\1\6\2\5\1\7\1\25"+
+    "\1\5\1\7\1\11\1\12\1\13\1\14\1\15\1\16"+
+    "\1\17\1\20\1\21\1\22\1\23\1\24\3\5\1\6"+
+    "\2\5\1\7\1\10\1\5\1\7\1\11\1\26\1\13"+
+    "\1\27\1\15\1\16\1\17\1\20\1\21\1\22\1\23"+
+    "\1\24\3\5\1\6\2\5\1\7\1\25\1\5\1\7"+
+    "\1\11\1\26\1\13\1\27\1\15\1\16\1\17\1\20"+
+    "\1\21\1\22\1\23\1\24\2\5\27\0\3\6\4\0"+
+    "\1\6\21\0\1\7\2\0\1\7\16\0\6\10\2\0"+
+    "\16\10\10\0\1\30\26\0\2\31\25\0\1\31\25\0"+
+    "\2\31\13\0\1\32\22\0\1\33\1\0\6\25\2\0"+
+    "\16\25\1\0\2\32\5\0\1\32\14\0\1\34\1\0"+
+    "\1\32\23\0\1\34";
 
   private static int [] zzUnpackTrans() {
-    int [] result = new int[308];
+    int [] result = new int[352];
     int offset = 0;
     offset = zzUnpackTrans(ZZ_TRANS_PACKED_0, offset, result);
     return result;
@@ -174,10 +180,10 @@ public class DependencyLanguageLexer implements FlexLexer {
   private static final int [] ZZ_ATTRIBUTE = zzUnpackAttribute();
 
   private static final String ZZ_ATTRIBUTE_PACKED_0 =
-    "\2\0\1\11\7\1\6\11\3\1\2\11\2\1\1\11";
+    "\4\0\1\11\10\1\6\11\4\1\2\11\2\1\1\11";
 
   private static int [] zzUnpackAttribute() {
-    int [] result = new int[24];
+    int [] result = new int[28];
     int offset = 0;
     offset = zzUnpackAttribute(ZZ_ATTRIBUTE_PACKED_0, offset, result);
     return result;
@@ -544,7 +550,7 @@ public class DependencyLanguageLexer implements FlexLexer {
             // fall through
           case 21: break;
           case 5: 
-            { return DependencyLanguageTypes.GREATER_THAN;
+            { return DependencyLanguageTypes.ANGLE_BRACKET_CLOSE;
             } 
             // fall through
           case 22: break;
@@ -554,7 +560,7 @@ public class DependencyLanguageLexer implements FlexLexer {
             // fall through
           case 23: break;
           case 7: 
-            { return DependencyLanguageTypes.LESS_THAN;
+            { return DependencyLanguageTypes.ANGLE_BRACKET_OPEN;
             } 
             // fall through
           case 24: break;
@@ -574,12 +580,12 @@ public class DependencyLanguageLexer implements FlexLexer {
             // fall through
           case 27: break;
           case 11: 
-            { return DependencyLanguageTypes.PARANTHESES_OPEN;
+            { yybegin(DEPENDENCY); return DependencyLanguageTypes.PARANTHESES_OPEN;
             } 
             // fall through
           case 28: break;
           case 12: 
-            { return DependencyLanguageTypes.PARANTHESES_CLOSE;
+            { yybegin(YYINITIAL); return DependencyLanguageTypes.PARANTHESES_CLOSE;
             } 
             // fall through
           case 29: break;